iStrategyInitializer syntax test pass

+++ ddd-main-domain/src/main/java/org/example/domain/strategy/repository/IStrategyRepo.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface IStrategyRepo {
 
-    List<StrategyAwardEntity> queryStrategyAwardEntityList(Integer strategyId);
+    List<StrategyAwardEntity> queryStrategyAwardEntityList(Long strategyId);
 }
