+++ ddd-main-domain/src/main/java/org/example/domain/strategy/service/IStrategyInitializer.java
@@ -2,5 +2,5 @@ package org.example.domain.strategy.service;
 
 public interface IStrategyInitializer {
 
-    void initializeStrategy(Integer strategyId);
+    void initializeStrategy(Long strategyId);
 }
